requirements.txt
streamlit
google-generativeai
openai-whisper
torch
numpy
